feat: Cell 내부 위치 조정 / NavigationBar 추가
