+++ HallofRhythm/HallofRhythm/GameViewController.swift
@@ -1,0 +1,17 @@
+//
+//  GameViewController.swift
+//  HallofRhythm
+//
+//  Created by HWANG-C-K on 2022/08/30.
+//
+
+import UIKit
+
+class GameViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        view.backgroundColor = .black
+        // Do any additional setup after loading the view.
+    }
+}
